risolto bug power-up

+++ Progetto/test/core/src/help/Info.java
@@ -19,7 +19,7 @@ public class Info {
     public static int hudHeight=80;
     public static int powerUpSpeed=3;
     public static float powerUpResize=0.5f;
-    public static int powerUpChance=2; //Probabilità da 1 a 10 che in un mattoncino ci sia un power up
+    public static float powerUpChance=2; //Probabilità da 1 a 10 che in un mattoncino ci sia un power up
     public static int numeroPowerUp=4; //Numero di power up esistenti (da aggiornare se se ne crea uno nuovo)
     public static int powerUpWidth=50;
     public static int powerUpHeight=45;
